refactor to speed up compile
