+++ bolt/exec/RowContainer.cpp
@@ -1821,92 +1821,6 @@ __attribute__((__visibility__("default"))) int8_t jit_GetDecodedIsNull(
       index);
 }
 
-__attribute__((__visibility__("default"))) void jit_HashAggrResizeVector(
-    char* vector,
-    int32_t size) {
-  reinterpret_cast<bytedance::bolt::BaseVector*>(vector)->resize(size);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatI8(
-    char* vector,
-    int32_t row,
-    int8_t value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<int8_t>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatI16(
-    char* vector,
-    int32_t row,
-    int16_t value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<int16_t>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatI32(
-    char* vector,
-    int32_t row,
-    int32_t value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<int32_t>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatI64(
-    char* vector,
-    int32_t row,
-    int64_t value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<int64_t>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatFloat(
-    char* vector,
-    int32_t row,
-    float value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<float>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetFlatDouble(
-    char* vector,
-    int32_t row,
-    double value,
-    int8_t isNull) {
-  auto* flat = reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-                   ->as<bytedance::bolt::FlatVector<double>>();
-  isNull ? flat->setNull(row, true) : flat->set(row, value);
-}
-
-__attribute__((__visibility__("default"))) void jit_HashAggrSetPartialAvgDouble(
-    char* vector,
-    int32_t row,
-    double sum,
-    int64_t count,
-    int8_t isNull) {
-  auto* rowVector =
-      reinterpret_cast<bytedance::bolt::BaseVector*>(vector)
-          ->as<bytedance::bolt::RowVector>();
-  auto* sumVector = rowVector->childAt(0)->asFlatVector<double>();
-  auto* countVector = rowVector->childAt(1)->asFlatVector<int64_t>();
-  if (isNull) {
-    rowVector->setNull(row, true);
-    return;
-  }
-  rowVector->setNull(row, false);
-  sumVector->set(row, sum);
-  countVector->set(row, count);
-}
-
 __attribute__((__visibility__("default"))) int8_t jit_ComplexTypeRowEqVectors(
     const char* row,
     int32_t offset,
