+++ bolt/functions/prestosql/aggregates/CountAggregate.cpp
@@ -91,96 +91,8 @@ class CountAggregate : public SimpleNumericAggregate<bool, int64_t, int64_t> {
         .scale = 0,
         .auxPrecision = 0,
         .auxScale = 0,
-        .ops = hashAggrJitOps()};
+        .ops = jit::getCountOps()};
   }
-
- private:
-  static void compileHashAggrJitInitGroup(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot) {
-    codegen.storeValue(
-        group,
-        codegen.builder().getInt64Ty(),
-        slot.offset,
-        codegen.builder().getInt64(0));
-  }
-
-  static void addInc(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot,
-      llvm::Value* inc) {
-    auto* state =
-        codegen.loadValue(group, codegen.builder().getInt64Ty(), slot.offset);
-    codegen.storeValue(
-        group,
-        codegen.builder().getInt64Ty(),
-        slot.offset,
-        codegen.builder().CreateAdd(state, inc));
-  }
-
-  static void compileHashAggrJitAddRawInput(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* /*decoded*/,
-      llvm::Value* /*row*/,
-      const jit::HashAggrJitSlot& slot,
-      bool,
-      llvm::BasicBlock*) {
-    addInc(codegen, group, slot, codegen.builder().getInt64(1));
-  }
-
-  static void compileHashAggrJitAddIntermediateResults(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool,
-      llvm::BasicBlock*) {
-    llvm::Value* inc = slot.desc.countStar
-        ? codegen.builder().getInt64(1)
-        : codegen.castValue(
-              codegen.loadDecodedValue(decoded, row, slot),
-              slot.desc.inputKind,
-              jit::HashAggrJitValueKind::Int64);
-    addInc(codegen, group, slot, inc);
-  }
-
-  static bool canCompileHashAggrJitExtract(
-      const jit::HashAggrJitSlot&,
-      bool) {
-    // count result is always BIGINT and never null.
-    return true;
-  }
-
-  static void compileHashAggrJitExtract(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot,
-      const jit::HashAggrJitExtractTarget& target) {
-    auto* value = codegen.loadValue(group, codegen.builder().getInt64Ty(), slot.offset);
-    codegen.emitFlatValue(
-        target.resultVector,
-        target.row,
-        jit::HashAggrJitValueKind::Int64,
-        value,
-        codegen.builder().getInt8(0));
-  }
-
-  static const jit::HashAggrJitOps* hashAggrJitOps() {
-    static const jit::HashAggrJitOps kOps{
-        "count",
-        &compileHashAggrJitInitGroup,
-        &compileHashAggrJitAddRawInput,
-        &compileHashAggrJitAddIntermediateResults,
-        &canCompileHashAggrJitExtract,
-        &compileHashAggrJitExtract};
-    return &kOps;
-  }
-
- public:
 #endif
 
   void toIntermediate(
