+++ bolt/functions/prestosql/aggregates/MinMaxAggregates.cpp
@@ -82,124 +82,7 @@ class MinMaxAggregate : public SimpleNumericAggregate<T, T, T> {
         .scale = 0,
         .auxPrecision = 0,
         .auxScale = 0,
-        .ops = hashAggrJitOps()};
-  }
-
- private:
-  static void compileHashAggrJitInitGroup(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot) {
-    codegen.setAccumulatorNull(group, slot);
-    auto* type = codegen.llvmType(slot.desc.accumulatorKind);
-    if (codegen.isFloatKind(slot.desc.accumulatorKind)) {
-      codegen.storeValue(group, type, slot.offset, llvm::ConstantFP::get(type, 0.0));
-    } else {
-      codegen.storeValue(group, type, slot.offset, llvm::ConstantInt::get(type, 0));
-    }
-  }
-
-  // min/max use the same logic for raw input and intermediate merge: both pick
-  // the better (min/max) of the decoded value and the current accumulator.
-  static void compileHashAggrJitUpdate(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool,
-      llvm::BasicBlock*) {
-    auto* value = codegen.castValue(
-        codegen.loadDecodedValue(decoded, row, slot),
-        slot.desc.inputKind,
-        slot.desc.accumulatorKind);
-    auto* type = codegen.llvmType(slot.desc.accumulatorKind);
-    auto* oldValue = codegen.loadValue(group, type, slot.offset);
-    auto* nullState = codegen.isAccumulatorNull(group, slot);
-    llvm::Value* better = nullptr;
-    if (codegen.isFloatKind(slot.desc.accumulatorKind)) {
-      auto* oldIsNan = codegen.builder().CreateFCmpUNO(oldValue, oldValue);
-      auto* valueIsNan = codegen.builder().CreateFCmpUNO(value, value);
-      if (slot.desc.kind == jit::HashAggrJitKind::Min) {
-        better = codegen.builder().CreateOr(
-            codegen.builder().CreateAnd(oldIsNan, codegen.builder().CreateNot(valueIsNan)),
-            codegen.builder().CreateAnd(
-                codegen.builder().CreateNot(valueIsNan),
-                codegen.builder().CreateFCmpOGT(oldValue, value)));
-      } else {
-        better = codegen.builder().CreateAnd(
-            codegen.builder().CreateNot(oldIsNan),
-            codegen.builder().CreateOr(
-                valueIsNan, codegen.builder().CreateFCmpOLT(oldValue, value)));
-      }
-    } else {
-      better = slot.desc.kind == jit::HashAggrJitKind::Min
-          ? codegen.builder().CreateICmpSLT(value, oldValue)
-          : codegen.builder().CreateICmpSGT(value, oldValue);
-    }
-    auto* shouldStore = codegen.builder().CreateOr(nullState, better);
-    codegen.storeValue(
-        group,
-        type,
-        slot.offset,
-        codegen.builder().CreateSelect(shouldStore, value, oldValue));
-    codegen.clearAccumulatorNull(group, slot);
-  }
-
-  static void compileHashAggrJitAddRawInput(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool checkInputNulls,
-      llvm::BasicBlock* nextBlock) {
-    compileHashAggrJitUpdate(
-        codegen, group, decoded, row, slot, checkInputNulls, nextBlock);
-  }
-
-  static void compileHashAggrJitAddIntermediateResults(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool checkInputNulls,
-      llvm::BasicBlock* nextBlock) {
-    compileHashAggrJitUpdate(
-        codegen, group, decoded, row, slot, checkInputNulls, nextBlock);
-  }
-
-  static bool canCompileHashAggrJitExtract(
-      const jit::HashAggrJitSlot& slot,
-      bool) {
-    // Flat setters exist for i8/i16/i32/i64/f32/f64 only. Int128 (long decimal)
-    // and Bool have no flat setter yet, fall back to non-JIT extract.
-    return slot.desc.accumulatorKind != jit::HashAggrJitValueKind::Int128 &&
-        slot.desc.accumulatorKind != jit::HashAggrJitValueKind::Bool;
-  }
-
-  static void compileHashAggrJitExtract(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot,
-      const jit::HashAggrJitExtractTarget& target) {
-    auto* value = codegen.loadValue(group, codegen.llvmType(slot.desc.accumulatorKind), slot.offset);
-    auto* isNull = codegen.builder().CreateZExt(
-        codegen.isAccumulatorNull(group, slot), codegen.builder().getInt8Ty());
-    codegen.emitFlatValue(
-        target.resultVector, target.row, slot.desc.accumulatorKind, value, isNull);
-  }
-
-  static const jit::HashAggrJitOps* hashAggrJitOps() {
-    static const jit::HashAggrJitOps kOps{
-        "minmax",
-        &compileHashAggrJitInitGroup,
-        &compileHashAggrJitAddRawInput,
-        &compileHashAggrJitAddIntermediateResults,
-        &canCompileHashAggrJitExtract,
-        &compileHashAggrJitExtract};
-    return &kOps;
+        .ops = jit::getMinMaxOps()};
   }
 
  protected:
