+++ bolt/functions/sparksql/aggregates/DecimalSumAggregate.h
@@ -99,126 +99,8 @@ class DecimalSumAggregate : public exec::Aggregate {
         .scale = resultScale,
         .auxPrecision = 0,
         .auxScale = 0,
-        .ops = hashAggrJitOps()};
+        .ops = jit::getDecimalSumOps()};
   }
-
- private:
-  static void compileHashAggrJitInitGroup(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot) {
-    codegen.setAccumulatorNull(group, slot);
-    codegen.builder().CreateCall(
-        codegen.module().getFunction("jit_HashAggrInitDecimalSum"),
-        {group, codegen.builder().getInt32(slot.offset)});
-  }
-
-  static void compileHashAggrJitAddRawInput(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool,
-      llvm::BasicBlock*) {
-    auto* rawValue = codegen.loadDecodedValue(decoded, row, slot);
-    codegen.clearAccumulatorNull(group, slot);
-    const auto helper = slot.desc.inputKind == jit::HashAggrJitValueKind::Int128
-        ? "jit_HashAggrUpdateDecimalSumI128"
-        : "jit_HashAggrUpdateDecimalSumI64";
-    codegen.builder().CreateCall(
-        codegen.module().getFunction(helper),
-        {group,
-         codegen.builder().getInt32(slot.offset),
-         slot.desc.inputKind == jit::HashAggrJitValueKind::Int128
-             ? codegen.castValue(
-                   rawValue, slot.desc.inputKind, jit::HashAggrJitValueKind::Int128)
-             : rawValue});
-  }
-
-  static void compileHashAggrJitAddIntermediateResults(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      llvm::Value* decoded,
-      llvm::Value* row,
-      const jit::HashAggrJitSlot& slot,
-      bool,
-      llvm::BasicBlock* nextBlock) {
-    auto* function = codegen.builder().GetInsertBlock()->getParent();
-    auto* continueBlock = llvm::BasicBlock::Create(
-        codegen.module().getContext(),
-        "sum_decimal_merge_cont",
-        function,
-        nextBlock);
-    auto* overflowBlock = llvm::BasicBlock::Create(
-        codegen.module().getContext(),
-        "sum_decimal_merge_overflow",
-        function,
-        continueBlock);
-    auto* mergeBlock = llvm::BasicBlock::Create(
-        codegen.module().getContext(),
-        "sum_decimal_merge",
-        function,
-        continueBlock);
-    auto* sumIsNull = codegen.isDecodedRowFieldNull(decoded, row, 0);
-    auto* isEmpty = codegen.loadDecodedRowField(
-        decoded, row, 1, jit::HashAggrJitValueKind::Int8);
-    auto* isNotEmpty =
-        codegen.builder().CreateICmpEQ(isEmpty, codegen.builder().getInt8(0));
-    auto* isOverflow = codegen.builder().CreateAnd(sumIsNull, isNotEmpty);
-    codegen.builder().CreateCondBr(isOverflow, overflowBlock, mergeBlock);
-
-    codegen.builder().SetInsertPoint(overflowBlock);
-    codegen.setAccumulatorNull(group, slot);
-    codegen.builder().CreateBr(continueBlock);
-
-    codegen.builder().SetInsertPoint(mergeBlock);
-    auto* sum = codegen.loadDecodedRowField(decoded, row, 0, slot.desc.inputKind);
-    codegen.clearAccumulatorNull(group, slot);
-    const auto helper = slot.desc.inputKind == jit::HashAggrJitValueKind::Int128
-        ? "jit_HashAggrMergeDecimalSumI128"
-        : "jit_HashAggrMergeDecimalSumI64";
-    codegen.builder().CreateCall(
-        codegen.module().getFunction(helper),
-        {group,
-         codegen.builder().getInt32(slot.offset),
-         slot.desc.inputKind == jit::HashAggrJitValueKind::Int128
-             ? codegen.castValue(
-                   sum, slot.desc.inputKind, jit::HashAggrJitValueKind::Int128)
-             : sum,
-         isEmpty});
-    codegen.builder().CreateBr(continueBlock);
-
-    codegen.builder().SetInsertPoint(continueBlock);
-  }
-
-  static bool canCompileHashAggrJitExtract(
-      const jit::HashAggrJitSlot&,
-      bool) {
-    return true;
-  }
-
-  static void compileHashAggrJitExtract(
-      jit::HashAggrJitCodegen& codegen,
-      llvm::Value* group,
-      const jit::HashAggrJitSlot& slot,
-      const jit::HashAggrJitExtractTarget& target) {
-    codegen.emitDecimalSumExtract(
-        target.resultVector, target.row, group, slot, target.partialOutput);
-  }
-
-  static const jit::HashAggrJitOps* hashAggrJitOps() {
-    static const jit::HashAggrJitOps kOps{
-        "sum_decimal",
-        &compileHashAggrJitInitGroup,
-        &compileHashAggrJitAddRawInput,
-        &compileHashAggrJitAddIntermediateResults,
-        &canCompileHashAggrJitExtract,
-        &compileHashAggrJitExtract};
-    return &kOps;
-  }
-
- public:
 #endif
 
   void initializeNewGroups(
